Fixed various issues with Registration/Login/AddProduct Forms
